Address https://github.com/digital-asset/daml/pull/3188#issuecomment-542616799
