+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/package.scala
@@ -33,7 +33,6 @@ package object tests {
     "TransactionServiceIT" -> (new TransactionService(_)),
     "WitnessesIT" -> (new Witnesses(_)),
     "WronglyTypedContractIdIT" -> (new WronglyTypedContractId(_)),
-    "CommandSubmissionLetIT" -> (new CommandSubmissionLet(_)),
     "LedgerConfigurationServiceIT" -> (new LedgerConfigurationService(_))
   )
